refactor: auto-split test_security.py
