+++ tests/unit/api/security/test_usage_tracker_contract.py
@@ -1,0 +1,33 @@
+"""Tests for security - Authentication and authorization utilities.
+
+These tests verify the security module using Design by Contract principles.
+"""
+
+import os
+from datetime import timedelta
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+# Configure async tests to use asyncio backend only
+pytestmark = pytest.mark.anyio
+
+
+@pytest.fixture(scope="module")
+def anyio_backend() -> str:
+    """Use asyncio backend only (trio not installed)."""
+    return "asyncio"
+
+
+class TestUsageTrackerContract:
+    """Design by Contract tests for UsageTracker class."""
+
+    def test_security_instantiates(self) -> None:
+        """Postcondition: UsageTracker can be instantiated."""
+        with patch.dict(
+            os.environ, {"GOLF_API_SECRET_KEY": "test-secret-key-32chars-long!!"}
+        ):
+            from src.api.auth.security import UsageTracker
+
+            tracker = UsageTracker()
+            assert tracker is not None
